fix final windows warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@38224 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 2f7da5fd500e293697c14260ecaea414461381b2

+++ Rec/Velo/VeloRecMonitors/src/PVResolution.cpp
@@ -273,7 +273,7 @@ StatusCode PVResolution::execute()
       m_odinTime0 = odin->gpsTime();
     }
     else {
-      unsigned long itime = odin->gpsTime() - m_odinTime0;
+      unsigned long itime = (unsigned long) (odin->gpsTime() - m_odinTime0);
       m_odinTime = 1e-6 * itime;
     }
